DCS-1192 code refactor Moved the getMoves() function from its own class into the model/prison/PrisonService class
