+++ src/main/kotlin/uk/gov/justice/digital/hmpps/welcometoprison/model/MovementService.kt
@@ -2,10 +2,11 @@ package uk.gov.justice.digital.hmpps.welcometoprison.model
 
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.welcometoprison.model.basm.BasmService
+import uk.gov.justice.digital.hmpps.welcometoprison.model.prison.PrisonService
 import java.time.LocalDate
 
 @Service
 class MovementService(val basmService: BasmService, val prisonService: PrisonService) {
   fun getMovements(agencyId: String, date: LocalDate) =
-    basmService.getMoves(agencyId, date, date) + (prisonService.getMoves(agencyId, date))
+    basmService.getMoves(agencyId, date, date) + prisonService.getMoves(agencyId, date)
 }
